+++ include/ignition/sensors/Lidar.hh
@@ -21,6 +21,7 @@
 #include <string>
 #include <vector>
 
+#include <ignition/common/SuppressWarning.hh>
 #include <ignition/common/Event.hh>
 
 #include "ignition/sensors/lidar/Export.hh"
@@ -70,6 +71,11 @@ namespace ignition
       /// \return true if the update was successfull
       public: virtual bool IGN_DEPRECATED(4) PublishLidarScan(
         const ignition::common::Time &_now);
+
+      /// \brief Apply noise to the laser buffer, if noise has been
+      /// configured. This should be called before PublishLidarScan if you
+      /// want the scan data to contain noise.
+      public: void ApplyNoise();
 
       /// \brief Publish LaserScan message
       /// \param[in] _now The current time
@@ -242,8 +248,10 @@ namespace ignition
       // Documentation inherited
       public: virtual bool IsActive() const;
 
+      IGN_COMMON_WARN_IGNORE__DLL_INTERFACE_MISSING
       /// \brief Just a mutex for thread safety
       public: mutable std::mutex lidarMutex;
+      IGN_COMMON_WARN_RESUME__DLL_INTERFACE_MISSING
 
       /// \brief Raw buffer of laser data.
       public: float *laserBuffer = nullptr;
@@ -263,9 +271,11 @@ namespace ignition
                   unsigned int _heighti, unsigned int _channels,
                   const std::string &/*_format*/)> _subscriber);
 
+      IGN_COMMON_WARN_IGNORE__DLL_INTERFACE_MISSING
       /// \brief Data pointer for private data
       /// \internal
       private: std::unique_ptr<LidarPrivate> dataPtr;
+      IGN_COMMON_WARN_RESUME__DLL_INTERFACE_MISSING
     };
     }
   }