+++ src/GpuLidarSensor.cc
@@ -14,12 +14,22 @@
  * limitations under the License.
  *
 */
+
+#ifdef _WIN32
+#pragma warning(push)
+#pragma warning(disable: 4005)
+#pragma warning(disable: 4251)
+#endif
 #include <ignition/msgs/pointcloud_packed.pb.h>
-#include <ignition/msgs/Utility.hh>
-#include <ignition/transport/Node.hh>
+#ifdef _WIN32
+#pragma warning(pop)
+#endif
 
 #include <ignition/common/Console.hh>
 #include <ignition/common/Profiler.hh>
+#include <ignition/msgs/Utility.hh>
+#include <ignition/transport/Node.hh>
+
 #include "ignition/sensors/GpuLidarSensor.hh"
 #include "ignition/sensors/SensorFactory.hh"
 
@@ -29,7 +39,8 @@ using namespace ignition::sensors;
 class ignition::sensors::GpuLidarSensorPrivate
 {
   /// \brief Fill the point cloud packed message
-  public: void FillPointCloudMsg();
+  /// \param[in] _laserBuffer Lidar data buffer.
+  public: void FillPointCloudMsg(const float *_laserBuffer);
 
   /// \brief Rendering camera
   public: ignition::rendering::GpuRaysPtr gpuRays;
@@ -238,6 +249,9 @@ bool GpuLidarSensor::Update(const std::chrono::steady_clock::duration &_now)
   /// \todo(anyone) It would be nice to remove this copy.
   this->dataPtr->gpuRays->Copy(this->laserBuffer);
 
+  // Apply noise before publishing the data.
+  this->ApplyNoise();
+
   this->PublishLidarScan(_now);
 
   if (this->dataPtr->pointPub.HasConnections())
@@ -247,7 +261,7 @@ bool GpuLidarSensor::Update(const std::chrono::steady_clock::duration &_now)
       msgs::Convert(_now);
     this->dataPtr->pointMsg.set_is_dense(true);
 
-    this->dataPtr->FillPointCloudMsg();
+    this->dataPtr->FillPointCloudMsg(this->laserBuffer);
 
     {
       this->AddSequence(this->dataPtr->pointMsg.mutable_header());
@@ -292,7 +306,7 @@ ignition::math::Angle GpuLidarSensor::VFOV() const
 }
 
 //////////////////////////////////////////////////
-void GpuLidarSensorPrivate::FillPointCloudMsg()
+void GpuLidarSensorPrivate::FillPointCloudMsg(const float *_laserBuffer)
 {
   IGN_PROFILE("GpuLidarSensorPrivate::FillPointCloudMsg");
   uint32_t width = this->pointMsg.width();
@@ -325,8 +339,8 @@ void GpuLidarSensorPrivate::FillPointCloudMsg()
     {
       // Index of current point, and the depth value at that point
       auto index = j * width * channels + i * channels;
-      float depth = this->gpuRays->Data()[index];
-      float intensity = this->gpuRays->Data()[index + 1];
+      float depth = _laserBuffer[index];
+      float intensity = _laserBuffer[index + 1];
       uint16_t ring = j;
 
       int fieldIndex = 0;