+++ test/integration/air_pressure_plugin.cc
@@ -99,8 +99,14 @@ sdf::ElementPtr AirPressureToSdfWithNoise(const std::string &_name,
     ->GetElement("sensor");
 }
 
+/// \brief Test air pressure sensor
 class AirPressureSensorTest: public testing::Test
 {
+  // Documentation inherited
+  protected: void SetUp() override
+  {
+    ignition::common::Console::SetVerbosity(4);
+  }
 };
 
 /////////////////////////////////////////////////
@@ -125,10 +131,9 @@ TEST_F(AirPressureSensorTest, CreateAirPressure)
 
   // create the sensor using sensor factory
   ignition::sensors::SensorFactory sf;
-  sf.AddPluginPaths(ignition::common::joinPaths(PROJECT_BUILD_PATH, "lib"));
   std::unique_ptr<ignition::sensors::AirPressureSensor> sensor =
       sf.CreateSensor<ignition::sensors::AirPressureSensor>(airPressureSdf);
-  EXPECT_TRUE(sensor != nullptr);
+  ASSERT_NE(nullptr, sensor);
 
   EXPECT_EQ(name, sensor->Name());
   EXPECT_EQ(topic, sensor->Topic());
@@ -137,7 +142,7 @@ TEST_F(AirPressureSensorTest, CreateAirPressure)
   std::unique_ptr<ignition::sensors::AirPressureSensor> sensorNoise =
       sf.CreateSensor<ignition::sensors::AirPressureSensor>(
           airPressureSdfNoise);
-  EXPECT_TRUE(sensorNoise != nullptr);
+  ASSERT_NE(nullptr, sensorNoise);
 
   EXPECT_EQ(name, sensorNoise->Name());
   EXPECT_EQ(topicNoise, sensorNoise->Topic());
@@ -167,14 +172,13 @@ TEST_F(AirPressureSensorTest, SensorReadings)
   // create the sensor using sensor factory
   // try creating without specifying the sensor type and then cast it
   ignition::sensors::SensorFactory sf;
-  sf.AddPluginPaths(ignition::common::joinPaths(PROJECT_BUILD_PATH, "lib"));
   std::unique_ptr<ignition::sensors::Sensor> s =
       sf.CreateSensor(airPressureSdf);
   std::unique_ptr<ignition::sensors::AirPressureSensor> sensor(
       dynamic_cast<ignition::sensors::AirPressureSensor *>(s.release()));
 
   // Make sure the above dynamic cast worked.
-  EXPECT_TRUE(sensor != nullptr);
+  ASSERT_NE(nullptr, sensor);
 
   std::unique_ptr<ignition::sensors::Sensor> sNoise =
       sf.CreateSensor(airPressureSdfNoise);
@@ -182,7 +186,7 @@ TEST_F(AirPressureSensorTest, SensorReadings)
       dynamic_cast<ignition::sensors::AirPressureSensor *>(sNoise.release()));
 
   // Make sure the above dynamic cast worked.
-  EXPECT_TRUE(sensorNoise != nullptr);
+  ASSERT_NE(nullptr, sensorNoise);
 
   // verify initial readings
   EXPECT_DOUBLE_EQ(0.0, sensor->ReferenceAltitude());
@@ -234,8 +238,6 @@ TEST_F(AirPressureSensorTest, Topic)
 
   // Factory
   ignition::sensors::SensorFactory factory;
-  factory.AddPluginPaths(ignition::common::joinPaths(PROJECT_BUILD_PATH,
-      "lib"));
 
   // Default topic
   {