+++ test/integration/imu_plugin.cc
@@ -57,8 +57,14 @@ sdf::ElementPtr ImuToSdf(const std::string &_name,
     ->GetElement("sensor");
 }
 
+/// \brief Test IMU sensor
 class ImuSensorTest: public testing::Test
 {
+  // Documentation inherited
+  protected: void SetUp() override
+  {
+    ignition::common::Console::SetVerbosity(4);
+  }
 };
 
 /////////////////////////////////////////////////
@@ -79,10 +85,9 @@ TEST_F(ImuSensorTest, CreateImu)
 
   // create the sensor using sensor factory
   ignition::sensors::SensorFactory sf;
-  sf.AddPluginPaths(ignition::common::joinPaths(PROJECT_BUILD_PATH, "lib"));
   std::unique_ptr<ignition::sensors::ImuSensor> sensor =
       sf.CreateSensor<ignition::sensors::ImuSensor>(imuSdf);
-  EXPECT_TRUE(sensor != nullptr);
+  ASSERT_NE(nullptr, sensor);
 
   EXPECT_EQ(name, sensor->Name());
   EXPECT_EQ(topic, sensor->Topic());
@@ -108,14 +113,13 @@ TEST_F(ImuSensorTest, SensorReadings)
   // create the sensor using sensor factory
   // try creating without specifying the sensor type and then cast it
   ignition::sensors::SensorFactory sf;
-  sf.AddPluginPaths(ignition::common::joinPaths(PROJECT_BUILD_PATH, "lib"));
   std::unique_ptr<ignition::sensors::Sensor> s =
       sf.CreateSensor(imuSdf);
   std::unique_ptr<ignition::sensors::ImuSensor> sensor(
       dynamic_cast<ignition::sensors::ImuSensor *>(s.release()));
 
   // Make sure the above dynamic cast worked.
-  EXPECT_TRUE(sensor != nullptr);
+  ASSERT_NE(nullptr, sensor);
 
   // subscribe to the topic
   WaitForMessageTestHelper<ignition::msgs::IMU> msgHelper(topic);
@@ -239,8 +243,6 @@ TEST_F(ImuSensorTest, Topic)
 
   // Factory
   ignition::sensors::SensorFactory factory;
-  factory.AddPluginPaths(ignition::common::joinPaths(PROJECT_BUILD_PATH,
-      "lib"));
 
   // Default topic
   {