+++ test/integration/logical_camera_plugin.cc
@@ -27,7 +27,15 @@
 #include <ignition/sensors/Export.hh>
 
 #include <ignition/math/Helpers.hh>
+#ifdef _WIN32
+#pragma warning(push)
+#pragma warning(disable: 4005)
+#pragma warning(disable: 4251)
+#endif
 #include <ignition/msgs.hh>
+#ifdef _WIN32
+#pragma warning(pop)
+#endif
 #include <ignition/transport/Node.hh>
 
 #include "test_config.h"  // NOLINT(build/include)
@@ -78,8 +86,14 @@ sdf::ElementPtr LogicalCameraToSdf(const std::string &_name,
     ->GetElement("sensor");
 }
 
+/// \brief Test logical camera sensor
 class LogicalCameraSensorTest: public testing::Test
 {
+  // Documentation inherited
+  protected: void SetUp() override
+  {
+    ignition::common::Console::SetVerbosity(4);
+  }
 };
 
 /////////////////////////////////////////////////
@@ -105,10 +119,9 @@ TEST_F(LogicalCameraSensorTest, CreateLogicalCamera)
 
   // create the sensor using sensor factory
   ignition::sensors::SensorFactory sf;
-  sf.AddPluginPaths(ignition::common::joinPaths(PROJECT_BUILD_PATH, "lib"));
   std::unique_ptr<ignition::sensors::LogicalCameraSensor> sensor =
       sf.CreateSensor<ignition::sensors::LogicalCameraSensor>(logicalCameraSdf);
-  EXPECT_TRUE(sensor != nullptr);
+  ASSERT_NE(nullptr, sensor);
 
   EXPECT_EQ(name, sensor->Name());
   EXPECT_EQ(topic, sensor->Topic());
@@ -144,14 +157,13 @@ TEST_F(LogicalCameraSensorTest, DetectBox)
   // create the sensor using sensor factory
   // try creating without specifying the sensor type and then cast it
   ignition::sensors::SensorFactory sf;
-  sf.AddPluginPaths(ignition::common::joinPaths(PROJECT_BUILD_PATH, "lib"));
   std::unique_ptr<ignition::sensors::Sensor> s =
       sf.CreateSensor(logicalCameraSdf);
   std::unique_ptr<ignition::sensors::LogicalCameraSensor> sensor(
       dynamic_cast<ignition::sensors::LogicalCameraSensor *>(s.release()));
 
   // Make sure the above dynamic cast worked.
-  EXPECT_TRUE(sensor != nullptr);
+  ASSERT_NE(nullptr, sensor);
 
   // verify initial image
   auto img = sensor->Image();
@@ -245,8 +257,6 @@ TEST_F(LogicalCameraSensorTest, Topic)
 
   // Factory
   ignition::sensors::SensorFactory factory;
-  factory.AddPluginPaths(ignition::common::joinPaths(PROJECT_BUILD_PATH,
-      "lib"));
 
   // Default topic
   {